feat: add mixins for density styles of all components (#18761)
Similarly to `mdc-density`, we add a mixin for all of the
density styles that will exist in `src/material`.

Additionally, mdc-typography and mdc-theming mixins are simplified
now that the `all-theme` mixins can control individual theming system
parts. This does not work for `@angular/material` typography due to
a cyclic dependency which we could eliminate if we require people
to _explicitly_ import the file where the `mat-core` mixin originates
from.